+++ examples/tfe/main.tf
@@ -42,11 +42,10 @@ data "azurerm_client_config" "current" {}
 locals {
   cluster_name_short = trimprefix(local.cluster_name, "${local.account_code}-")
 
-  azuread_clusterrole_map = {
-    cluster_admin_users  = local.cluster_admin_users
-    cluster_view_users   = {}
-    standard_view_users  = {}
-    standard_view_groups = {}
+  rbac_bindings = {
+    cluster_admin_users = local.cluster_admin_users
+    cluster_view_users  = {}
+    cluster_view_groups = []
   }
 
   node_groups = {
@@ -266,7 +265,7 @@ module "aks" {
 
   admin_group_object_ids = [var.aad_group_id]
 
-  azuread_clusterrole_map = local.azuread_clusterrole_map
+  rbac_bindings = local.rbac_bindings
 
   node_groups = local.node_groups